requirements.txt
requests
fastapi
uvicorn
python-dotenv
icalendar
